requirements.txt
notion-client
openai
pandas
python-dotenv